fix: widen render index buffers from u16 to u32
A Volcano planet at depth 8 has 43,746 triangles, needing render-vertex
indices up to 131,237 — u16 caps out at 65,535. mesh_render_indices
silently wrapped (wrong triangles drawn); mesh_render_line_indices
panicked outright on the u16 addition overflow, since the wireframe
line-index buffer is always rebuilt in set_mesh regardless of whether
wireframe mode is active. This was unreachable before this feature,
since the app previously always hardcoded a fixed low depth with no UI
to pick a larger one. Widen mesh_render_indices/mesh_render_line_indices/
pack_index_buffer to u32 and switch the pipeline to
wgpu::IndexFormat::Uint32; new BDD scenarios cover a mesh large enough
to exceed u16's range for both functions.

+++ planet-renderer/src/gpu/buffers.rs
@@ -35,16 +35,16 @@ pub fn mesh_render_vertices(mesh: &Mesh, colors: &[Rgb]) -> Vec<Vertex> {
         .collect()
 }
 
-pub fn mesh_render_indices(mesh: &Mesh) -> Vec<u16> {
+pub fn mesh_render_indices(mesh: &Mesh) -> Vec<u32> {
     (0..3 * mesh.triangles().len())
-        .map(|index| index as u16)
+        .map(|index| index as u32)
         .collect()
 }
 
-pub fn mesh_render_line_indices(mesh: &Mesh) -> Vec<u16> {
+pub fn mesh_render_line_indices(mesh: &Mesh) -> Vec<u32> {
     (0..mesh.triangles().len())
         .flat_map(|i| {
-            let base = 3 * i as u16;
+            let base = (3 * i) as u32;
             [base, base + 1, base + 1, base + 2, base + 2, base]
         })
         .collect()
@@ -65,7 +65,7 @@ pub fn pack_vertex_buffer(vertices: &[Vertex]) -> Vec<u8> {
     bytes
 }
 
-pub fn pack_index_buffer(indices: &[u16]) -> Vec<u8> {
+pub fn pack_index_buffer(indices: &[u32]) -> Vec<u8> {
     let mut bytes = Vec::with_capacity(std::mem::size_of_val(indices));
     for index in indices {
         bytes.extend_from_slice(&index.to_le_bytes());
